refactor(server): move server files
