+++ server/.gitignore
@@ -1,1 +1,7 @@
+# Python
+__pycache__/
+*.py[cod]
+*$py.class
+
+# virtualenv
 venv
