+++ server/src/emoji/routes/api/history.py
