+++ server/src/emoji/routes/emoji.py
