+++ server/src/emoji/routes/healthcheck.py
