+++ server/src/emoji/routes/static.py
